Rename GNB to MainHeader and MainNavigation to GNB
