+++ components/layout/GNB.tsx
@@ -1,31 +1,20 @@
-import Link from "next/link";
-import { useState } from "react";
-
+import { BOOKMARKS_PATH, STUDIES_PATH, TEMPLATES_PATH } from "@/utils/routes";
 import styles from "@/styles/components/MainHeader.module.sass";
-import { HOME_PATH } from "@/utils/routes";
-import ProfileBtn from "../common/ProfileBtn";
-import ProfileSubmenu from "../common/ProfileSubmenu";
-import Logo from "../common/icon/Logo";
-import MainNavigation from "./MainNavigation";
+import Navigation from "../common/Navigation";
 
 function GNB() {
-  const [profileSubmenuVisibility, setProfileSubmenuVisibility] =
-    useState(false);
-  const handleProfileBtnClick = () => {
-    setProfileSubmenuVisibility(prev => !prev);
-  };
+  const gnbItems = [
+    { label: "스터디", route: STUDIES_PATH },
+    { label: "템플릿", route: TEMPLATES_PATH },
+    { label: "북마크", route: BOOKMARKS_PATH },
+  ];
 
   return (
-    <header className={styles.mainHeader}>
-      <div>
-        <Link href={HOME_PATH}>
-          <Logo className={styles.homeLogo} />
-        </Link>
-      </div>
-      <MainNavigation />
-      <ProfileBtn handleClick={handleProfileBtnClick} />
-      {profileSubmenuVisibility ? <ProfileSubmenu /> : null}
-    </header>
+    <Navigation aria-label="global-navigation-bar" className={styles.gnb}>
+      {gnbItems.map(item => (
+        <Navigation.Item route={item.route}>{item.label}</Navigation.Item>
+      ))}
+    </Navigation>
   );
 }
